refactor: 불필요 코드 정리, 리네이밍 #45
