--- src/main/java/org/sopt/app/application/auth/EncryptService.java
@@ -1,19 +1,0 @@
-package org.sopt.app.application.auth;
-
-import org.springframework.security.crypto.bcrypt.BCrypt;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class EncryptService implements PasswordEncoder {
-
-    @Override
-    public String encode(CharSequence rawPassword) {
-        return BCrypt.hashpw(rawPassword.toString(), BCrypt.gensalt());
-    }
-
-    @Override
-    public boolean matches(CharSequence rawPassword, String encodedPassword) {
-        return BCrypt.checkpw(rawPassword.toString(), encodedPassword);
-    }
-}
