--- src/main/java/org/sopt/app/application/user/UserInfoMapper.java
@@ -1,15 +1,0 @@
-package org.sopt.app.application.user;
-
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-
-@Mapper(
-        componentModel = "spring",
-        injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-        unmappedTargetPolicy = ReportingPolicy.ERROR
-)
-public interface UserInfoMapper {
-
-    UserInfo.Id of(Long id);
-}
