--- src/main/java/org/sopt/app/application/user/dto/LogInUserDto.java
@@ -1,10 +1,0 @@
-package org.sopt.app.application.user.dto;
-
-
-import lombok.Builder;
-
-public record LogInUserDto(String email, String password) {
-    @Builder
-    public LogInUserDto {
-    }
-}
