--- src/main/java/org/sopt/app/presentation/auth/dto/ChangeNicknameRequestDto.java
@@ -1,16 +1,0 @@
-package org.sopt.app.presentation.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ChangeNicknameRequestDto {
-
-  private String nickname;
-
-}
