--- src/main/java/org/sopt/app/presentation/auth/dto/ChangePasswordRequestDto.java
@@ -1,14 +1,0 @@
-package org.sopt.app.presentation.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ChangePasswordRequestDto {
-    String password;
-}
