--- src/main/java/org/sopt/app/presentation/user/request/CreateUserRequest.java
@@ -1,17 +1,0 @@
-package org.sopt.app.presentation.user.request;
-
-import com.sun.istack.NotNull;
-import lombok.Getter;
-import org.sopt.app.domain.enums.OsType;
-
-@Getter
-public class CreateUserRequest {
-    @NotNull
-    private String nickname;
-    @NotNull
-    private String email;
-    @NotNull
-    private String password;
-    private OsType osType;
-    private String clientToken;
-}
