--- src/main/java/org/sopt/app/presentation/user/response/LoginResponse.java
@@ -1,16 +1,0 @@
-package org.sopt.app.presentation.user.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class LoginResponse {
-    private Long userId;
-    private String profileMessage;
-
-    @Builder
-    public LoginResponse(Long userId, String profileMessage) {
-        this.userId = userId;
-        this.profileMessage = profileMessage;
-    }
-}
