--- src/main/java/org/sopt/app/presentation/user/response/SignUpResponse.java
@@ -1,15 +1,0 @@
-package org.sopt.app.presentation.user.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Builder
-@Setter
-public class SignUpResponse {
-
-  private Long userId;
-
-
-}
